fix: delete portfolio shopify

+++ app/Actions/Dropshipping/Shopify/Product/DeleteShopifyUserHasProduct.php
@@ -20,7 +20,7 @@ class DeleteShopifyUserHasProduct extends OrgAction
     use WithAttributes;
     use WithActionUpdate;
 
-    public function handle(?ShopifyUserHasProduct $product): int|null
+    public function handle(?ShopifyUserHasProduct $product, $forceDelete = false): ShopifyUserHasProduct|null|int
     {
         if (!$product) {
             return null;
@@ -29,6 +29,10 @@ class DeleteShopifyUserHasProduct extends OrgAction
         $shopifyUser = $product->shopifyUser;
         $shopifyUser->api()->getRestClient()->request('DELETE', '/admin/api/2025-04/products/'.$product->shopify_product_id.'.json');
 
-        return $product->delete();
+        if ($forceDelete) {
+            return $product->delete();
+        }
+
+        return $this->update($product, ['shopify_product_id' => null]);
     }
 }
